+++ include/solvers/fluid_dynamics_matrix_free.h
@@ -38,6 +38,9 @@ namespace dealii
   class MGCoarseGridApplyPreconditioner;
 }
 
+template <int dim, typename MGNumber>
+class MyMGTransferMatrixFree;
+
 /**
  * @brief A geometric multigrid preconditioner compatible with the
  * matrix-free solver.
@@ -64,7 +67,7 @@ protected:
   using VectorType         = LinearAlgebra::distributed::Vector<Number>;
   using MGVectorType       = LinearAlgebra::distributed::Vector<MGNumber>;
   using TrilinosVectorType = LinearAlgebra::distributed::Vector<double>;
-  using LSTransferType     = MGTransferMatrixFree<dim, MGNumber>;
+  using LSTransferType     = MyMGTransferMatrixFree<dim, MGNumber>;
   using GCTransferType     = MGTransferGlobalCoarsening<dim, MGVectorType>;
   using OperatorType       = NavierStokesOperatorBase<dim, MGNumber>;
   using SmootherPreconditionerType = PreconditionBase<MGVectorType>;
@@ -230,7 +233,7 @@ protected:
 
   /// Collection of boundary constraints and refinement edge constrations for
   /// the different levels in the local smoothing approach.
-  MGConstrainedDoFs mg_constrained_dofs;
+  MGLevelObject<MGConstrainedDoFs> mg_constrained_dofs;
 
   /// Transfer operator for local smoothing
   std::shared_ptr<LSTransferType> mg_transfer_ls;
@@ -270,6 +273,10 @@ protected:
   std::shared_ptr<PreconditionMG<dim, MGVectorType, GCTransferType>>
     gc_multigrid_preconditioner_intermediate;
 
+  /// Local smoothing multigrid preconditioner object (intermediate level)
+  std::shared_ptr<PreconditionMG<dim, MGVectorType, LSTransferType>>
+    ls_multigrid_preconditioner_intermediate;
+
   /// Multigrid method
   std::shared_ptr<Multigrid<MGVectorType>> mg;
 
@@ -291,7 +298,7 @@ protected:
   const DoFHandler<dim> &dof_handler;
 
   /// Describes the layout of DoFs for FE_Q_iso_Q1 elements.
-  const DoFHandler<dim> &dof_handler_fe_q_iso_q1;
+  DoFHandler<dim> dof_handler_fe_q_iso_q1;
 
   /// Vector holding number of coarse grid iterations
   mutable std::vector<unsigned int> coarse_grid_iterations;